requirements.txt
asyncpg==0.27.0
discord.py==2.1.0
openai==0.25.0
python-decouple==3.6
pytube==12.1.2
requests==2.28.1
SQLAlchemy==1.4.46
colorama==0.4.6
